api/error: moved to types/errors
